+++ lib/provider/editorconfig.ts
@@ -1,17 +1,12 @@
-/// <reference path="../typings/editorconfig.d.ts" />
-
-"use strict";
-
 import * as ts from "typescript";
 
 import * as editorconfig from "editorconfig";
 
-import {Options} from "../";
+import { Options } from "../";
 
 let emitBaseDirWarning = false;
 
 export default function makeFormatCodeOptions(fileName: string, opts: Options, formatOptions: ts.FormatCodeOptions): Promise<ts.FormatCodeOptions> {
-    "use strict";
 
     if (opts.verbose && opts.baseDir && !emitBaseDirWarning) {
         console.log("editorconfig is not supported baseDir options");